move store and refine useNdk and useSigner
