+++ src/hooks/use-signer/index.ts
@@ -1,0 +1,10 @@
+import { useStore } from 'src/store';
+
+export const useSigner = () => {
+  const ndk = useStore((state) => state.ndk);
+  const setSigner = useStore((state) => state.setSigner);
+
+  const signer = ndk.signer;
+
+  return { signer, setSigner };
+};
